blueprint: Update blueprint based on our demo
This commit updates the MEAN blueprint to be broken down in the
same way as the MEAN application in https://github.com/quilt/demos.
The demos repo did a better job of decomposing the functionality
into different files in the canonical manner for Quilt, and these
two repos are essentially identical, so we should delete the demos
repo after merging this commit.

This commit copies the files from the demos repository and re-names
the mean application from TodoApp (used in the demos repo) to Mean.